phase1/requirements.txt
Flask
redis
python-dotenv
requests
